move Layout to layouts folder
