+++ src/layouts/Layout.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from './Navbar'
+import Navbar from '../components/Navbar'
 import { Outlet } from 'react-router-dom'
 
 const Layout = () => {
